requirements.txt
pandas==2.2.2
requests==2.32.3
python-dotenv==1.0.1
google-cloud-bigquery==3.25.0
pyarrow==17.0.0
dbt-bigquery==1.8.1
pytest==8.2.2

gspread
google-auth-oauthlib
